Add option-passing to plugin interface, refactor go generator into two plugins
