--- plugins/go/openapi_go_client/templates/app.yaml.tmpl
@@ -1,9 +1,0 @@
-application: {{.Renderer.Name}}
-version: 1
-runtime: go
-api_version: go1
-handlers:
-- url: /.*
-  script: _go_app
-- url: /
-  static_dir: static
