+++ plugins/go/openapi_go_client/templates/client.go.tmpl
@@ -1,4 +1,4 @@
-package {{.Package}}
+package {{.Renderer.Package}}
 
 import (
   "encoding/json"
