+++ plugins/go/openapi_go_client/templates/types.go.tmpl
@@ -1,4 +1,4 @@
-package {{.Package}}
+package {{.Renderer.Package}}
 
 import (
 )
