+++ plugins/go/openapi_go_service/templates/app.go.tmpl
@@ -1,4 +1,4 @@
-package {{.Package}}
+package {{.Renderer.Package}}
 
 import (
 	"encoding/json"
