+++ plugins/go/openapi_go_service/templates/service.go.tmpl
@@ -1,4 +1,4 @@
-package {{.Package}}
+package {{.Renderer.Package}}
 
 import (
 	"log"
