+++ plugins/go/openapi_go_service/templates/types.go.tmpl
@@ -1,4 +1,4 @@
-package {{.Package}}
+package {{.Renderer.Package}}
 
 import (
 )
